# docs/swagger.yaml
definitions:
  ApiError:
    properties:
      error:
        example: Validation error
        type: string
    type: object
  Classroom:
    properties:
      desc:
        example: 25 | Филиалы Челябинский филиал | Лекционная
        type: string
      id:
        example: "3587"
        type: string
      title:
        example: "24"
        type: string
    type: object
  Group:
    properties:
      desc:
        example: Финансовый факультет | Очная
        type: string
      id:
        example: "110694"
        type: string
      title:
        example: ФФ21-1
        type: string
    type: object
  LoginRequest:
    properties:
      password:
        example: Bibt8877
        type: string
      username:
        example: "227789"
        type: string
    required:
    - password
    - username
    type: object
  LoginResponse:
    properties:
      token:
        example: eyJhbGciOiJSUzI1NiIsInR5cCI6IkpXVCJ9
        type: string
    type: object
  MiniScheduleItem:
    properties:
      date:
        example: 2023.11.27
        type: string
      lessonType:
        example: Лекции
        type: string
    type: object
  ScheduleItem:
    properties:
      address:
        example: ул. Олеко Дундича, 23
        type: string
      classroom:
        example: ОД/341
        type: string
      date:
        example: 2023.11.27
        type: string
      endsAt:
        example: "11:40"
        type: string
      lecturer:
        example: Бердышев Александр Валентинович
        type: string
      lesson:
        example: Иностранный язык
        type: string
      lessonNumberEnd:
        example: 1
        type: integer
      lessonNumberStart:
        example: 1
        type: integer
      lessonType:
        example: Лекции
        type: string
      startsAt:
        example: "10:10"
        type: string
      weekDay:
        example: 0
        type: integer
    type: object
  Teacher:
    properties:
      desc:
        example: Департамент гуманитарных наук
        type: string
      id:
        example: 00000000-0001-2345-6789-000000005451
        type: string
      title:
        example: Махаматов Таир
        type: string
    type: object
externalDocs:
  description: OpenAPI
  url: https://swagger.io/resources/open-api/
host: localhost:5051/
info:
  contact:
    email: support@swagger.io
    name: https://web.telegram.org/a/#488960669
    url: http://www.swagger.io/support
  description: This is a server finapp-api.
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  termsOfService: http://swagger.io/terms/
  title: Finapp-api
  version: "1.0"
paths:
  /auth/login:
    post:
      description: In success case returns Access JWT Token
      parameters:
      - description: Credentials input
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/LoginRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/LoginResponse'
            type: array
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/ApiError'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/ApiError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/ApiError'
      summary: Try to sign in user
      tags:
      - auth
  /classrooms/:
    get:
      description: Return list of classrooms which found by provided query term
      parameters:
      - description: Classroom search mask
        in: query
        name: term
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/Classroom'
            type: array
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/ApiError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/ApiError'
      summary: Return List of classrooms
      tags:
      - classrooms
  /groups/:
    get:
      description: Return list of student groups which found by provided query term
      parameters:
      - description: Group search mask
        in: query
        name: term
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/Group'
            type: array
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/ApiError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/ApiError'
      summary: Return List of groups
      tags:
      - groups
  /schedule/classroom:
    get:
      description: Returns schedule for provided classroom Id and time interval
      parameters:
      - description: Classroom ID
        in: query
        name: id
        required: true
        type: string
      - description: Start date
        in: query
        name: from
        required: true
        type: string
      - description: End date
        in: query
        name: to
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/ScheduleItem'
            type: array
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/ApiError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/ApiError'
      summary: Return schedule for provided classroom
      tags:
      - schedule
  /schedule/classroom/mini:
    get:
      description: Returns schedule in compact format for provided classroom Id and
        time interval
      parameters:
      - description: Classroom ID
        in: query
        name: id
        required: true
        type: string
      - description: Start date
        in: query
        name: from
        required: true
        type: string
      - description: End date
        in: query
        name: to
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/MiniScheduleItem'
            type: array
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/ApiError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/ApiError'
      summary: Return mini schedule for provided classroom
      tags:
      - schedule
  /schedule/group:
    get:
      description: Returns schedule for provided group Id and time interval
      parameters:
      - description: Group ID
        in: query
        name: id
        required: true
        type: string
      - description: Start date
        in: query
        name: from
        required: true
        type: string
      - description: End date
        in: query
        name: to
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/ScheduleItem'
            type: array
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/ApiError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/ApiError'
      summary: Return schedule for provided group
      tags:
      - schedule
  /schedule/group/mini:
    get:
      description: Returns schedule in compact format for provided teacher Id and
        time interval
      parameters:
      - description: Teacher ID
        in: query
        name: id
        required: true
        type: string
      - description: Start date
        in: query
        name: from
        required: true
        type: string
      - description: End date
        in: query
        name: to
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/MiniScheduleItem'
            type: array
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/ApiError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/ApiError'
      summary: Return mini schedule for provided group
      tags:
      - schedule
  /schedule/teacher:
    get:
      description: Returns schedule for provided teacher Id and time interval
      parameters:
      - description: Teacher ID
        in: query
        name: id
        required: true
        type: string
      - description: Start date
        in: query
        name: from
        required: true
        type: string
      - description: End date
        in: query
        name: to
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/ScheduleItem'
            type: array
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/ApiError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/ApiError'
      summary: Return schedule for provided teacher
      tags:
      - schedule
  /schedule/teacher/mini:
    get:
      description: Returns schedule in compact format for provided teacher Id and
        time interval
      parameters:
      - description: Teacher ID
        in: query
        name: id
        required: true
        type: string
      - description: Start date
        in: query
        name: from
        required: true
        type: string
      - description: End date
        in: query
        name: to
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/MiniScheduleItem'
            type: array
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/ApiError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/ApiError'
      summary: Return mini schedule for provided teacher
      tags:
      - schedule
  /teachers/:
    get:
      description: Return list of teachers which found by provided query term
      parameters:
      - description: Teacher search mask
        in: query
        name: term
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/Teacher'
            type: array
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/ApiError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/ApiError'
      summary: Return List of teachers
      tags:
      - teachers
securityDefinitions:
  BasicAuth:
    type: basic
swagger: "2.0"