# internal/dto/classroom.go
package dto

type Classroom struct {
	Id          string `json:"id" example:"3587"`
	Title       string `json:"title" example:"24"`
	Description string `json:"desc" example:"25 | Филиалы Челябинский филиал | Лекционная"`
} // @name Classroom